+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/commands.py
@@ -9,9 +9,10 @@ from azure.cli.core.util import empty_on_404
 
 from ._client_factory import cf_container_services
 from ._client_factory import cf_managed_clusters
-from ._format import aks_get_versions_table_format
 from ._format import aks_list_table_format
 from ._format import aks_show_table_format
+from ._format import aks_upgrades_table_format
+from ._format import aks_versions_table_format
 
 
 def load_command_table(self, _):
@@ -57,7 +58,7 @@ def load_command_table(self, _):
         g.custom_command('create', 'aks_create', no_wait_param='no_wait')
         g.command('delete', 'delete', no_wait_param='raw', confirmation=True)
         g.custom_command('get-credentials', 'aks_get_credentials')
-        g.command('get-versions', 'get_upgrade_profile', table_transformer=aks_get_versions_table_format)
+        g.command('get-upgrades', 'get_upgrade_profile', table_transformer=aks_upgrades_table_format)
         g.custom_command('install-cli', 'k8s_install_cli', client_factory=None)
         g.custom_command('install-connector', 'k8s_install_connector')
         g.custom_command('list', 'aks_list', table_transformer=aks_list_table_format)
@@ -68,3 +69,7 @@ def load_command_table(self, _):
                          confirmation='Kubernetes may be unavailable during cluster upgrades.\n' +
                          'Are you sure you want to perform this operation?')
         g.generic_wait_command('wait')
+
+    with self.command_group('aks', container_services_sdk, client_factory=cf_container_services) as g:
+        g.custom_command('get-versions', 'aks_get_versions', table_transformer=aks_versions_table_format,
+                         deprecate_info="az aks get-upgrades")