+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/tests/latest/test_storage_blob_scenarios.py
@@ -89,7 +89,7 @@ class StorageBlobUploadTests(StorageScenarioMixin, ScenarioTest):
              JMESPathCheck('properties.contentLength', file_size_kb * 1024)])
 
         self.storage_cmd('storage blob service-properties show', account_info) \
-            .assert_with_checks(JMESPathCheck('hourMetrics.enabled', False))
+            .assert_with_checks(JMESPathCheck('hourMetrics.enabled', True))
 
         if not skip_download:
             downloaded = os.path.join(local_dir, 'test.file')
@@ -278,6 +278,41 @@ class StorageBlobUploadTests(StorageScenarioMixin, ScenarioTest):
             .assert_with_checks(JMESPathCheck('deleted', True))
         self.storage_cmd('storage container exists -n {}', account_info, c) \
             .assert_with_checks(JMESPathCheck('exists', False))
+
+    @ResourceGroupPreparer()
+    @StorageAccountPreparer()
+    def test_storage_blob_soft_delete(self, resource_group, storage_account):
+        account_info = self.get_account_info(resource_group, storage_account)
+        container = self.create_container(account_info)
+
+        # create a blob
+        local_file = self.create_temp_file(1)
+        blob_name = self.create_random_name(prefix='blob', length=24)
+
+        self.storage_cmd('storage blob upload -c {} -f "{}" -n {} --type block', account_info,
+                         container, local_file, blob_name)
+        self.assertEqual(len(self.storage_cmd('storage blob list -c {}',
+                                              account_info, container).get_output_in_json()), 1)
+
+        # set delete-policy to enable soft-delete
+        self.storage_cmd('storage blob service-properties delete-policy update --enable true --days-retained 2',
+                         account_info)
+        self.storage_cmd('storage blob service-properties delete-policy show',
+                         account_info).assert_with_checks(JMESPathCheck('enabled', True),
+                                                          JMESPathCheck('days', 2))
+
+        # soft-delete and check
+        self.storage_cmd('storage blob delete -c {} -n {}', account_info, container, blob_name)
+        self.assertEqual(len(self.storage_cmd('storage blob list -c {}',
+                                              account_info, container).get_output_in_json()), 0)
+
+        self.assertEqual(len(self.storage_cmd('storage blob list -c {} --include d',
+                                              account_info, container).get_output_in_json()), 1)
+
+        # undelete and check
+        self.storage_cmd('storage blob undelete -c {} -n {}', account_info, container, blob_name)
+        self.assertEqual(len(self.storage_cmd('storage blob list -c {}',
+                                              account_info, container).get_output_in_json()), 1)
 
 
 if __name__ == '__main__':