+++ src/command_modules/azure-cli-cognitiveservices/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+0.1.11
+++++++
+* Update the 'notice' of creating a new Cognitive Service account.
+
 0.1.10
 ++++++
 * Update for CLI core changes.